Added some colors to the css file and made links work
