+++ src/about/about.component.html
@@ -5,6 +5,16 @@
     </header>
 
     <div class="navigation-buttons">
+        <a
+            routerLink="/manage"
+            class="nav-button manage-colors-button"
+            aria-label="Navigate to Color Management"
+        >
+            <div class="button-content">
+                <h3>Manage Colors</h3>
+            </div>
+        </a>
+
         <a 
             routerLink="/home"
             class="nav-button home-button"
@@ -22,16 +32,6 @@
         >
             <div class="button-content">
                 <h3>Color Coordinate</h3>
-            </div>
-        </a>
-
-        <a
-            routerLink="/manage_colors"
-            class="nav-button manage-colors-button"
-            aria-label="Navigate to Color Management"
-        >
-            <div class="button-content">
-                <h3>Manage Colors</h3>
             </div>
         </a>
     </div>
