+++ src/color/color_coordinate_generation.component.html
@@ -8,6 +8,16 @@
     </header>
 
     <div class="navigation-buttons">
+      <a
+            routerLink="/manage"
+            class="nav-button manage-colors-button"
+            aria-label="Navigate to Color Management"
+        >
+            <div class="button-content">
+                <h3>Manage Colors</h3>
+            </div>
+        </a>
+        
         <a 
             routerLink="/home"
             class="nav-button home-button"
@@ -25,16 +35,6 @@
         >
             <div class="button-content">
                 <h3>About</h3>
-            </div>
-        </a>
-
-        <a
-            routerLink="/manage_colors"
-            class="nav-button manage-colors-button"
-            aria-label="Navigate to Color Management"
-        >
-            <div class="button-content">
-                <h3>Manage Colors</h3>
             </div>
         </a>
     </div>
