+++ src/home/home.component.html
@@ -25,7 +25,7 @@
         </a>
 
         <a
-            routerLink="/manage_colors"
+            routerLink="/manage"
             class="nav-button manage-colors-button"
             aria-label="Navigate to Color Management"
         >
