+++ src/manage/manage_colors.component.html
@@ -1,0 +1,139 @@
+<div class="color-management-container">
+  
+  <div class="navigation-buttons">
+    <a 
+    routerLink="/about"
+    class="nav-button about-button"
+    aria-label="Navigate to About Page"
+    >
+    <div class="button-content">
+      <h3>About Our Team</h3>
+    </div>
+    </a>
+    <a 
+    routerLink="/home"
+    class="nav-button home-button"
+    aria-label="Navigate to Home Page"
+    >
+    <div class="button-content">
+      <h3>Home</h3>
+    </div>
+    </a>
+    <a
+    routerLink="/color"
+    class="nav-button color-button"
+    aria-label="Navigate to Color Coordinate Generator"
+    >
+    <div class="button-content">
+      <h3>Color Coordinate</h3>
+    </div>
+    </a>
+  </div>
+      
+  <header class="color-management-header">
+    <h1>Color Management</h1>
+  </header>
+
+  <div class="tab-buttons">
+    <button (click)="switchTab('add')" [class.active]="activeTab === 'add'">Add Color</button>
+    <button (click)="switchTab('edit')" [class.active]="activeTab === 'edit'">Edit Color</button>
+    <button (click)="switchTab('delete')" [class.active]="activeTab === 'delete'">Delete Color</button>
+  </div>
+
+  <div *ngIf="errorMessage" class="alert error">
+    {{ errorMessage }}
+  </div>
+  <div *ngIf="successMessage" class="alert success">
+    {{ successMessage }}
+  </div>
+
+  <div *ngIf="isLoading" class="loading">
+    Loading colors...
+  </div>
+
+  <div *ngIf="activeTab === 'add'" class="tab-content">
+    <h2>Add New Color</h2>
+    <form [formGroup]="addForm" (ngSubmit)="onAddSubmit()">
+      <div class="form-group">
+        <label for="add-name">Color Name</label>
+        <input id="add-name" type="text" formControlName="name" placeholder="e.g. Sky Blue">
+      </div>
+      
+      <div class="form-group">
+        <label for="add-hex">Hex Value</label>
+        <input id="add-hex" type="text" formControlName="hexValue" placeholder="#RRGGBB">
+        <div class="color-preview" [style.background]="addForm.value.hexValue || '#fff'"></div>
+      </div>
+      
+      <button type="submit" [disabled]="addForm.invalid">Add Color</button>
+    </form>
+  </div>
+  
+  <div *ngIf="activeTab === 'edit'" class="tab-content">
+    <h2>Edit Existing Color</h2>
+    <form [formGroup]="editForm" (ngSubmit)="onEditSubmit()">
+      <div class="form-group">
+        <label for="edit-select">Select Color to Edit</label>
+        <select id="edit-select" formControlName="selectedColor" (change)="onColorSelectForEdit(editForm.value.selectedColor)">
+          <option *ngFor="let color of colors" [value]="color.id">
+            {{ color.name }} ({{ color.hexValue }})
+          </option>
+        </select>
+      </div>
+      
+      <div class="form-group">
+        <label for="edit-name">New Name</label>
+        <input id="edit-name" type="text" formControlName="name">
+      </div>
+      
+      <div class="form-group">
+        <label for="edit-hex">New Hex Value</label>
+        <input id="edit-hex" type="text" formControlName="hexValue">
+        <div class="color-preview" [style.background]="editForm.value.hexValue || '#fff'"></div>
+      </div>
+      
+      <button type="submit" [disabled]="editForm.invalid">Update Color</button>
+    </form>
+  </div>
+
+  <div *ngIf="activeTab === 'delete'" class="tab-content">
+    <h2>Delete Color</h2>
+    <form [formGroup]="deleteForm" (ngSubmit)="onDeleteSubmit()">
+      <div class="form-group">
+        <label for="delete-select">Select Color to Delete</label>
+        <select id="delete-select" formControlName="selectedColor">
+          <option *ngFor="let color of colors" [value]="color.id">
+            {{ color.name }} ({{ color.hexValue }})
+          </option>
+        </select>
+      </div>
+      
+      <div class="warning" *ngIf="colors.length <= 2">
+        Cannot delete color - you must maintain at least 2 colors.
+      </div>
+      
+      <button type="submit" [disabled]="deleteForm.invalid || colors.length <= 2">Delete Color</button>
+    </form>
+  </div>
+
+  <div class="logo-container">
+    <img src="https://raw.githubusercontent.com/HridayaB/hridaya-app/master/src/assets/images/group_project_logo.png" alt="Project Logo" class="logo" />
+  </div>
+
+  <div class="accessibility">
+    <h2>Color Accessibility</h2>
+    <p>All color combinations meet WCAG AA and WCAG AAA requirements:</p>
+    
+    <div class="contrast-links">
+      <a *ngFor="let contrast of config.contrastLinks" 
+         [href]="contrast.url" 
+         target="_blank" 
+         rel="noopener noreferrer"
+         class="contrast-chip"
+         [style.background]="contrast.bgColor"
+         [style.color]="contrast.textColor">
+         {{ contrast.textColor }} on {{ contrast.bgColor }} ({{ contrast.ratio }}:1)
+      </a>
+    </div>
+  </div>
+</div>
