+++ src/manage/manage_colors.config.ts
@@ -1,0 +1,52 @@
+export const colorManageConfig = {
+    contrastLinks: [
+        {
+          textColor: '#32492C',
+          bgColor: '#D5DEC9',
+          ratio: '7.11',
+          url: 'https://webaim.org/resources/contrastchecker/?fcolor=32492C&bcolor=D5DEC9'
+        },
+        {
+          textColor: '#443D6C',
+          bgColor: '#D5DEC9',
+          ratio: '7.09',
+          url: 'https://webaim.org/resources/contrastchecker/?fcolor=443D6C&bcolor=D5DEC9'
+        },
+        {
+          textColor: '#244A4C',
+          bgColor: '#D5DEC9',
+          ratio: '7',
+          url: 'https://webaim.org/resources/contrastchecker/?fcolor=244A4C&bcolor=D5DEC9'
+        },
+        {
+          textColor: '#1B3B3B',
+          bgColor: '#D5DEC9',
+          ratio: '8.71',
+          url: 'https://webaim.org/resources/contrastchecker/?fcolor=1B3B3B&bcolor=D5DEC9'
+        },
+        {
+          textColor: '#8A0000',
+          bgColor: '#D5DEC9',
+          ratio: '7.26',
+          url: 'https://webaim.org/resources/contrastchecker/?fcolor=8A0000&bcolor=D5DEC9'
+        },
+        {
+            textColor: '#6B0000',
+            bgColor: '#D5DEC9',
+            ratio: '9.3',
+            url: 'https://webaim.org/resources/contrastchecker/?fcolor=6B0000&bcolor=D5DEC9'
+        },
+        {
+            textColor: '#004D00',
+            bgColor: '#D5DEC9',
+            ratio: '7.33',
+            url: 'https://webaim.org/resources/contrastchecker/?fcolor=004D00&bcolor=D5DEC9'
+        },
+        {
+            textColor: '#613A00',
+            bgColor: '#D5DEC9',
+            ratio: '7.15',
+            url: 'https://webaim.org/resources/contrastchecker/?fcolor=613A00&bcolor=D5DEC9'
+        }
+    ]
+};
